+++ spec/spec_helper.rb
@@ -36,19 +36,6 @@ module EY
           old_info(*args)
         end
       end
-
-      def logged_system(cmd)
-        output = `#{cmd} 2>&1`
-        successful = ($? == 0)
-        if ENV['VERBOSE']
-          if successful
-            $stdout.puts "#{cmd}\n#{output.strip}".chomp
-          else
-            $stderr.puts "\nCommand `#{cmd}` exited with status #{$?.exitstatus}: '#{output.strip}'"
-          end
-        end
-        successful
-      end
     end
 
     class Strategies::Git