+++ ChangeLog.md
@@ -2,7 +2,7 @@
 
 ## NEXT
 
-  *
+  * Replace Open4 with systemu
 
 ## v1.5.34 (2012-03-22)